+++ lib/honeylixir_tracing/application.ex
@@ -3,9 +3,9 @@ defmodule HoneylixirTracing.Application do
 
   use Application
 
-  # @known_integrations %{plug: []}
-
   def start(_type, _args) do
+    ecto_integration()
+
     Supervisor.start_link(children(), strategy: :one_for_one)
   end
 
@@ -28,6 +28,29 @@ defmodule HoneylixirTracing.Application do
 
       _ ->
         []
+    end
+  end
+
+  defp ecto_integration() do
+    case Map.get(integration_configs(), :ecto) do
+      true ->
+        raise """
+          HoneylixirTracing cannot accept only `true` for this integration.
+          You must provide either :repo_name or :event_prefix as a config value.
+          You can configure it by doing
+          config :honeylixir_tracing, integrations: %{ecto: [repo_name: :awesome_app]}
+        """
+
+      opts when is_list(opts) ->
+        HoneylixirTracing.Integrations.Ecto.setup(opts)
+
+      opts when not is_nil(opts) ->
+        raise """
+          Please provide a Keyword list of configuration as documented for this integration.
+        """
+
+      _ ->
+        nil
     end
   end